esp udp support, refactor debug functions.
