+++ src/test/esp/esp_server.c
@@ -40,31 +40,7 @@ static int esp_mode = 0;
 static int use_udp = 0;
 
 /* esp security association info */
-static uint8_t in_sa_gcm[ESP_SPI_LEN] = {0x01, 0x01, 0x01, 0x01};
-static uint8_t out_sa_gcm[ESP_SPI_LEN] = {0x02, 0x02, 0x02, 0x02};
-static uint8_t in_sa_cbc[ESP_SPI_LEN] = {0x03, 0x03, 0x03, 0x03};
-static uint8_t out_sa_cbc[ESP_SPI_LEN] = {0x04, 0x04, 0x04, 0x04};
-static uint8_t in_sa_des3[ESP_SPI_LEN] = {0x05, 0x05, 0x05, 0x05};
-static uint8_t out_sa_des3[ESP_SPI_LEN] = {0x06, 0x06, 0x06, 0x06};
-/* 32 byte key + 4 byte nonce*/
-static uint8_t in_enc_key[36] =
-     {0x03, 0x03, 0x03, 0x03, 0x03, 0x03, 0x03, 0x03,
-      0x03, 0x03, 0x03, 0x03, 0x03, 0x03, 0x03, 0x03,
-      0x03, 0x03, 0x03, 0x03, 0x03, 0x03, 0x03, 0x03,
-      0x03, 0x03, 0x03, 0x03, 0x03, 0x03, 0x03, 0x03,
-      0x0a, 0x0b, 0x0c, 0x0d};
-static uint8_t out_enc_key[36] =
-     {0x04, 0x04, 0x04, 0x04, 0x04, 0x04, 0x04, 0x04,
-      0x04, 0x04, 0x04, 0x04, 0x04, 0x04, 0x04, 0x04,
-      0x04, 0x04, 0x04, 0x04, 0x04, 0x04, 0x04, 0x04,
-      0x04, 0x04, 0x04, 0x04, 0x04, 0x04, 0x04, 0x04,
-      0x0a, 0x0b, 0x0c, 0x0d};
-static uint8_t in_auth_key[16] =
-     {0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
-      0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01};
-static uint8_t out_auth_key[16] =
-     {0x02, 0x02, 0x02, 0x02, 0x02, 0x02, 0x02, 0x02,
-      0x02, 0x02, 0x02, 0x02, 0x02, 0x02, 0x02, 0x02};
+#include "esp_common.c"
 
 int main(int argc, char * argv[])
 {
