+++ src/test/esp/test_esp.c
@@ -56,31 +56,8 @@ static int client_connected = 0;
 static const uint8_t test_pattern[16] = {0x54, 0x65, 0x73, 0x74, 0x20, 0x70,
                                          0x61, 0x74, 0x74, 0x65, 0x72, 0x6e,
                                          0x20, 0x2d, 0x20, 0x2d};
-static uint8_t in_sa_gcm[ESP_SPI_LEN] = {0x01, 0x01, 0x01, 0x01};
-static uint8_t out_sa_gcm[ESP_SPI_LEN] = {0x02, 0x02, 0x02, 0x02};
-static uint8_t in_sa_cbc[ESP_SPI_LEN] = {0x03, 0x03, 0x03, 0x03};
-static uint8_t out_sa_cbc[ESP_SPI_LEN] = {0x04, 0x04, 0x04, 0x04};
-static uint8_t in_sa_des3[ESP_SPI_LEN] = {0x05, 0x05, 0x05, 0x05};
-static uint8_t out_sa_des3[ESP_SPI_LEN] = {0x06, 0x06, 0x06, 0x06};
-/* 32 byte key + 4 byte nonce*/
-static uint8_t in_enc_key[36] =
-     {0x03, 0x03, 0x03, 0x03, 0x03, 0x03, 0x03, 0x03,
-      0x03, 0x03, 0x03, 0x03, 0x03, 0x03, 0x03, 0x03,
-      0x03, 0x03, 0x03, 0x03, 0x03, 0x03, 0x03, 0x03,
-      0x03, 0x03, 0x03, 0x03, 0x03, 0x03, 0x03, 0x03,
-      0x0a, 0x0b, 0x0c, 0x0d};
-static uint8_t out_enc_key[36] =
-     {0x04, 0x04, 0x04, 0x04, 0x04, 0x04, 0x04, 0x04,
-      0x04, 0x04, 0x04, 0x04, 0x04, 0x04, 0x04, 0x04,
-      0x04, 0x04, 0x04, 0x04, 0x04, 0x04, 0x04, 0x04,
-      0x04, 0x04, 0x04, 0x04, 0x04, 0x04, 0x04, 0x04,
-      0x0a, 0x0b, 0x0c, 0x0d};
-static uint8_t in_auth_key[16] =
-     {0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01,
-      0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01, 0x01};
-static uint8_t out_auth_key[16] =
-     {0x02, 0x02, 0x02, 0x02, 0x02, 0x02, 0x02, 0x02,
-      0x02, 0x02, 0x02, 0x02, 0x02, 0x02, 0x02, 0x02};
+/* esp security association info */
+#include "esp_common.c"
 
 /* wolfIP: server side callback. */
 static void server_cb(int fd, uint16_t event, void *arg)
