+++ tools/ip-xfrm/esp_sa.txt
@@ -2,9 +2,9 @@
 # You can add (but not modify or delete) records with the command line option:
 # -o 'uat:esp_sa:"protocol","srcIP","dstIP","spi","encryption_algo","encryption_key_string","authentication_algo","authentication_key_string","sn_length","sn_upper"'
 #"Protocol","Src IP","Dest IP","SPI","Encryption","Encryption Key","Authentication","Authentication Key","SN","ESN High Bits"
-"IPv4","10.10.10.1","10.10.10.2","0x05050505","NULL","","HMAC-SHA-256-128 [RFC4868]","0x01010101010101010101010101010101","32-bit","0"
-"IPv4","10.10.10.2","10.10.10.1","0x06060606","NULL","","HMAC-SHA-256-128 [RFC4868]","0x02020202020202020202020202020202","32-bit","0"
 "IPv4","10.10.10.1","10.10.10.2","0x03030303","AES-CBC [RFC3602]","0x0303030303030303030303030303030303030303030303030303030303030303","HMAC-SHA-256-128 [RFC4868]","0x01010101010101010101010101010101","32-bit","0"
 "IPv4","10.10.10.2","10.10.10.1","0x04040404","AES-CBC [RFC3602]","0x0404040404040404040404040404040404040404040404040404040404040404","HMAC-SHA-256-128 [RFC4868]","0x02020202020202020202020202020202","32-bit","0"
 "IPv4","10.10.10.1","10.10.10.2","0x01010101","AES-GCM with 16 octet ICV [RFC4106]","0x03030303030303030303030303030303030303030303030303030303030303030a0b0c0d","NULL","","32-bit","0"
 "IPv4","10.10.10.2","10.10.10.1","0x02020202","AES-GCM with 16 octet ICV [RFC4106]","0x04040404040404040404040404040404040404040404040404040404040404040a0b0c0d","NULL","","32-bit","0"
+"IPv4","10.10.10.1","10.10.10.2","0x05050505","TripleDES-CBC [RFC2451]","0x030303030303030303030303030303030303030303030303","HMAC-SHA-256-128 [RFC4868]","0x01010101010101010101010101010101","32-bit","0"
+"IPv4","10.10.10.2","10.10.10.1","0x06060606","TripleDES-CBC [RFC2451]","0x040404040404040404040404040404040404040404040404","HMAC-SHA-256-128 [RFC4868]","0x02020202020202020202020202020202","32-bit","0"
